plat: rpi5: scmi: add device resources management policy
Add RPI5 device resources management policy. Each device is described by
struct scmi_device which allows to define resets and pinctrl resources used
by device (resources). All Device's resources allowed to be accessed by
Agent when the Agent is given access to the Device.

The access to the Device resource expected to be denied if Agent is
not given permission to access this resource.

The current patch implements permission checking only for SCMI reset
domains for now.

NOTE. Shared access to the Device resources is expected to be managed by
corresponding platform drivers (including, but not limited to, reference
counters).

Signed-off-by: Grygorii Strashko <grygorii_strashko@epam.com>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>