plat: rpi5: platform.mk: scmi build fix
The build has option to enable SCMI SCMI_SERVER_SUPPORT, but in
RPI5 platfrom.mk the SCMI is partially under if (SCMI_SERVER_SUPPORT)
statement and partially not which cause build to fail when
SCMI_SERVER_SUPPORT is not defined or set to 0.

Move all SCMI specific code under SCMI_SERVER_SUPPORT==1 to avoid build
failures.

Signed-off-by: Grygorii Strashko <grygorii_strashko@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>